Display and minor fixes
- search box rendering wrong fixed (search icon is slightly off center
now, will have to find another way to fix that)
- updated material UI rating import to fit with material UI v5
- atempted to fix 400 error when navigating with search box. No success

"so if u start in the oceania region and go straight to anyplace in the
americas without going to a diff region first (asia africa europe etc)
it will give a 400 error.
so if u wanna go straight to LA u hv to go to some random place in
africa europe or asia first, AND THEN go to LA to avoid geting a 400
error.
this is what ive learnt from trail and error lol, i hv no idea how to
fix it or why this error occurs.
and after u get the 400 error its permanent. u hv to reload the page to
get rid of it"